Fusion: Change FusionObjectID to u32.

+++ lib/fusion/object.h
@@ -44,7 +44,7 @@ typedef bool (*FusionPropIterator)( char *key, void *value, void *ctx);
 
 
 
-typedef unsigned long FusionObjectID;
+typedef u32 FusionObjectID;
 
 
 typedef enum {
